+++ src/MShop.Application/UseCases/Product/DeleteProduct/DeleteProduct.cs
@@ -4,6 +4,7 @@ using MShop.Core.Base;
 using MShop.Core.Message;
 using MShop.Repository.Interface;
 using MShop.Core.Data;
+using MShop.Core.DomainObject;
 
 namespace MShop.Application.UseCases.Product.DeleteProduct
 {
@@ -26,17 +27,25 @@ namespace MShop.Application.UseCases.Product.DeleteProduct
             _unitOfWork = unitOfWork;
         }
 
-        public async Task<ProductModelOutPut> Handle(DeleteProductInPut request, CancellationToken cancellationToken)
+        public async Task<Result<ProductModelOutPut>> Handle(DeleteProductInPut request, CancellationToken cancellationToken)
         {
             var product = await _productRespository.GetById(request.Id);
-            NotifyExceptionIfNull(product, "Não foi possivel localizar a produto da base de dados!");
+            //NotifyExceptionIfNull(product, "Não foi possivel localizar a produto da base de dados!");
+            if (NotifyErrorIfNull(product, "Não foi possivel localizar a produto da base de dados!"))
+                return Result<ProductModelOutPut>.Error();
 
             var hasImages = await _imageRepository.Filter(x => x.ProductId == product!.Id);
             if(hasImages?.Count > 0)
-                NotifyException("Existe(m) Imagen(s) associada(s) a esse produto");
+            {
+                Notify("Existe(m) Imagen(s) associada(s) a esse produto");
+                return Result<ProductModelOutPut>.Error();
+            }
+                
 
             product!.ProductRemovedEvent();
-            NotifyExceptionIfNull(product.Events.Count == 0 ? null : product.Events, $" Não foi possivel registrar o event ProductDeletedEvent");
+            //NotifyExceptionIfNull(product.Events.Count == 0 ? null : product.Events, $" Não foi possivel registrar o event ProductDeletedEvent");
+            if (NotifyErrorIfNull(product.Events.Count == 0 ? null : product.Events, $" Não foi possivel registrar o event ProductDeletedEvent"))
+                return Result<ProductModelOutPut>.Error();
 
             await _productRespository.DeleteById(product, cancellationToken);
             await _unitOfWork.CommitAsync(cancellationToken);   
@@ -53,7 +62,7 @@ namespace MShop.Application.UseCases.Product.DeleteProduct
                 product.IsActive, 
                 product.CategoryId);*/
 
-            return ProductModelOutPut.FromProduct(product); 
+            return Result<ProductModelOutPut>.Success(ProductModelOutPut.FromProduct(product)); 
         }
     }
 }
